use aggregate in queries, reorganize shadcn stuff, more auth

+++ src/app/artpiece/[id]/reviews/actions/delete_review.ts
@@ -1,5 +1,6 @@
 "use server";
 
+import { getUser } from "@/app/lib/auth/user";
 import sql from "@/app/lib/sql";
 import { redirect, RedirectType } from "next/navigation";
 
@@ -7,6 +8,11 @@ export default async function delete_review(
   userId: number,
   artpieceId: number,
 ) {
+  if (!(await getUser())) {
+    console.error("Non-user trying to delete a review");
+    return;
+  }
+
   await sql`
     DELETE FROM review
     WHERE user_id = ${userId} AND artpiece_id = ${artpieceId}
